workload-identity-python/src/requirements.txt
azure-identity
azure-keyvault-secrets